requirements.txt
markdownify
requests
# Core data processing and analysis
pandas
numpy
openpyxl

# Development environment
jupyterlab
python-dotenv

# LangChain framework and extensions
langchain
langchain-core
langchain-community
langchain_chroma
langchain-mistralai
langchain-anthropic
langgraph
langgraph-checkpoint
langgraph-checkpoint-sqlite
langchain_tavily

# LLM providers and API clients
openai
anthropic
mistralai
litellm==1.65.4

# observers
langfuse
smolagents[telemetry]
opentelemetry-sdk
opentelemetry-exporter-otlp
openinference-instrumentation-smolagents

# Vector databases and indexing
llama-index

# Agent frameworks
smolagents
smolagents[telemetry]
# UI components
gradio
chainlit

# Search and tools
tavily-python
helium
selenium
arxiv
wikipedia

# graph drawing
grandalf